requirements.txt
Django==1.5
gdata==2.0.17
gunicorn==0.17.2
psycopg2==2.4.6
wsgiref==0.1.2
django-user-accounts